+++ python/turbodbc/connect.py
@@ -16,7 +16,7 @@ def _make_connection_string(dsn, **kwargs):
 
 
 @translate_exceptions
-def connect(dsn=None, read_buffer_size=None, rows_to_buffer=None, parameter_sets_to_buffer=None, use_async_io=False, prefer_unicode=False, **kwargs):
+def connect(dsn=None, read_buffer_size=None, parameter_sets_to_buffer=None, use_async_io=False, prefer_unicode=False, **kwargs):
     """
     Create a connection with the database identified by the dsn
     :param dsn: data source name as given in the odbc.ini file
@@ -38,12 +38,6 @@ def connect(dsn=None, read_buffer_size=None, rows_to_buffer=None, parameter_sets
                       would like to connect with (e.g. `user` or `password`)
     """    
     connection = Connection(intern_connect(_make_connection_string(dsn, **kwargs)))
-
-    if rows_to_buffer:
-        warnings.warn("Calling turbodbc.connect() with parameter rows_to_buffer is deprecated. "
-                      "Instead, set the parameter read_buffer_size to turbodbc.Megabytes(x) or "
-                      "turbodbc.Rows(y) instead.", DeprecationWarning)
-        connection.impl.set_buffer_size(Rows(rows_to_buffer))
 
     if read_buffer_size:
         connection.impl.set_buffer_size(read_buffer_size)