+++ python/turbodbc_test/test_connect.py
@@ -55,12 +55,10 @@ def test_connect_raises_on_invalid_additional_option(dsn, configuration):
 @for_one_database
 def test_connect_performance_settings(dsn, configuration):
     connection = connect(dsn=dsn,
-                         rows_to_buffer=317,
                          parameter_sets_to_buffer=123,
                          use_async_io=True,
                          **get_credentials(configuration))
 
-    assert connection.impl.get_buffer_size().rows == 317
     assert connection.impl.parameter_sets_to_buffer == 123
     assert connection.impl.use_async_io == True
 
@@ -74,12 +72,6 @@ def test_connect_with_rows(dsn, configuration):
 @for_one_database
 def test_connect_with_megabytes(dsn, configuration):
     connection = connect(dsn=dsn, read_buffer_size=Megabytes(1), **get_credentials(configuration))
-    assert connection.impl.get_buffer_size().megabytes == 1
-
-
-@for_one_database
-def test_read_buffer_size_has_priority_to_rows_to_buffer(dsn, configuration):
-    connection = connect(dsn=dsn, read_buffer_size=Megabytes(1), rows_to_buffer=317, **get_credentials(configuration))
     assert connection.impl.get_buffer_size().megabytes == 1